+++ pytket/tests/classical_test.py
@@ -1044,3 +1044,37 @@ def test_arithmetic_ops() -> None:
     assert str(commands[1].op.get_exp()) == "(b << 2)"
     assert str(commands[2].op.get_exp()) == "(c >> 2)"
     assert str(commands[3].op.get_exp()) == "((a ** c) - b)"
+
+
+def test_renaming() -> None:
+    circ = Circuit()
+    a = circ.add_c_register("a", 3)
+    b = circ.add_c_register("b", 3)
+    c = circ.add_c_register("c", 3)
+    circ.add_classicalexpbox_bit(a[0] & b[0] | c[0], [a[0]])
+    circ.add_classicalexpbox_bit(a[0] & c[2], [c[0]])
+    d = [Bit("d", index) for index in range(0, 3)]
+    bmap = {a[0]: d[0], b[0]: d[1], c[0]: d[2]}
+    original_commands = circ.get_commands()
+    assert circ.rename_units(bmap)
+    commands = circ.get_commands()
+    assert str(commands[0].op.get_exp()) == "((d[0] & d[1]) | d[2])"
+    assert commands[0].args == [bmap[arg] for arg in original_commands[0].args]
+    assert str(commands[1].op.get_exp()) == "(d[0] & c[2])"
+    assert commands[1].args == [
+        bmap[arg] if arg in bmap else arg for arg in original_commands[1].args
+    ]
+    assert DecomposeClassicalExp().apply(circ)
+
+    # Register-wise renaming should raise error
+    circ = Circuit()
+    a = circ.add_c_register("a", 3)
+    b = circ.add_c_register("b", 3)
+    c = circ.add_c_register("c", 3)
+    circ.add_classicalexpbox_register(a + b // c, a)
+    bmap = {a[0]: d[0]}
+
+    with pytest.raises(ValueError) as e:
+        circ.rename_units(bmap)
+    err_msg = f"Can't rename bits in {a.__repr__()}"
+    assert err_msg in str(e.value)